+++ docs/ARCHITECTURE.md
@@ -87,7 +87,7 @@ lib/
 - **`api/`** — Dio-based HTTP clients. Each client is its own class exposed through a `Provider`. Auth (Twitch OAuth for IGDB, Bearer for TMDB / SteamGridDB, etc.) lives inside the client. Errors are normalized into custom `*ApiException` types via `api_error_extract.dart`.
 - **`database/`** — `DatabaseService` (singleton via `databaseServiceProvider`) delegates CRUD to DAOs. Schema is declared in `schema.dart` (`DatabaseSchema.create*Table`). Migrations are incremental: one file per version under `migrations/`, registered in `MigrationRegistry.all`. **Never edit historical migration SQL** — add a new version instead.
 - **`services/`** — anything that's not an API client and not raw CRUD: `.xcoll` / `.xcollx` import/export, tracker sync (RA, Steam, Trakt, Kodi, MAL, AniList), Discord RPC, update checker, image cache, config.
-- **`import/`** — library import layer (ports & adapters). The `ImportSource` port with one adapter per source under `sources/<name>/`, plus shared domain pieces (`ImportWriter` for collection write, `TmdbMatcher`, `RateLimitedRetry`). **Kinorium** is the first adapter on it; the other source importers still live in `services/*_import_service.dart` and migrate onto the layer one at a time. See [core/import/README.md](../lib/core/import/README.md).
+- **`import/`** — library import layer (ports & adapters). The `ImportSource` port with one adapter per source under `sources/<name>/`, plus shared domain pieces (`ImportWriter` for collection write, `TmdbMatcher`, `RateLimitedRetry`). **Kinorium, Steam, RetroAchievements, MyAnimeList and AniList** are all adapters on the layer, each implementing `ImportSource` and writing through `ImportWriter`. `trakt_zip_import_service.dart` in `services/` is the last importer still carrying its own copy of this logic. See [core/import/README.md](../lib/core/import/README.md).
 - **`logging/`** — wrapper around `package:logging`. `print` is forbidden in production code.
 
 ### `data/`
