--- test/core/services/steam_import_service_test.dart
@@ -1,706 +1,0 @@
-import 'package:flutter_test/flutter_test.dart';
-import 'package:mocktail/mocktail.dart';
-import 'package:tonkatsu_box/core/api/steam_api.dart';
-import 'package:tonkatsu_box/core/services/steam_import_service.dart';
-import 'package:tonkatsu_box/shared/models/collection_item.dart';
-import 'package:tonkatsu_box/shared/models/game.dart';
-import 'package:tonkatsu_box/shared/models/item_status.dart';
-import 'package:tonkatsu_box/shared/models/media_type.dart';
-import 'package:tonkatsu_box/shared/models/wishlist_item.dart';
-
-import '../../helpers/test_helpers.dart';
-
-void main() {
-  setUpAll(registerAllFallbacks);
-
-  late SteamImportService sut;
-  late MockSteamApi mockSteamApi;
-  late MockIgdbApi mockIgdbApi;
-  late MockDatabaseService mockDb;
-  late MockGameDao mockGameDao;
-  late MockWishlistDao mockWishlistDao;
-
-  setUp(() {
-    mockSteamApi = MockSteamApi();
-    mockIgdbApi = MockIgdbApi();
-    mockDb = MockDatabaseService();
-    mockGameDao = MockGameDao();
-    when(() => mockDb.gameDao).thenReturn(mockGameDao);
-    mockWishlistDao = MockWishlistDao();
-    when(() => mockDb.wishlistDao).thenReturn(mockWishlistDao);
-
-    sut = SteamImportService(
-      steamApi: mockSteamApi,
-      igdbApi: mockIgdbApi,
-      database: mockDb,
-    );
-  });
-
-  void setupStandardMocks({
-    List<SteamOwnedGame>? library,
-    int collectionId = 1,
-  }) {
-    final List<SteamOwnedGame> games = library ??
-        <SteamOwnedGame>[
-          createTestSteamOwnedGame(
-            appId: 440,
-            name: 'Team Fortress 2',
-            playtimeMinutes: 1250,
-            lastPlayed: DateTime(2024, 1, 28),
-          ),
-          createTestSteamOwnedGame(
-            appId: 570,
-            name: 'Dota 2',
-            playtimeMinutes: 0,
-          ),
-        ];
-
-    when(() => mockSteamApi.getOwnedGames(
-          apiKey: any(named: 'apiKey'),
-          steamId: any(named: 'steamId'),
-        )).thenAnswer((_) async => games);
-
-    when(() => mockIgdbApi.lookupSteamGames(any()))
-        .thenAnswer((Invocation inv) async {
-      final List<String> appIds =
-          inv.positionalArguments[0] as List<String>;
-      final Map<String, Game> result = <String, Game>{};
-      for (final String appId in appIds) {
-        result[appId] = Game(id: appId.hashCode.abs(), name: 'Game $appId');
-      }
-      return result;
-    });
-
-    when(() => mockDb.findCollectionItem(
-          collectionId: any(named: 'collectionId'),
-          mediaType: any(named: 'mediaType'),
-          externalId: any(named: 'externalId'),
-        )).thenAnswer((_) async => null);
-
-    when(() => mockGameDao.upsertGame(any())).thenAnswer((_) async {});
-
-    when(() => mockDb.addItemToCollection(
-          collectionId: any(named: 'collectionId'),
-          mediaType: any(named: 'mediaType'),
-          externalId: any(named: 'externalId'),
-          platformId: any(named: 'platformId'),
-          status: any(named: 'status'),
-        )).thenAnswer((_) async => 100);
-
-    when(() => mockDb.updateItemTimeSpent(any(), any()))
-        .thenAnswer((_) async {});
-
-    when(() => mockDb.updateItemActivityDates(
-          any(),
-          lastActivityAt: any(named: 'lastActivityAt'),
-        )).thenAnswer((_) async {});
-
-    when(() => mockWishlistDao.addWishlistItem(
-          text: any(named: 'text'),
-          mediaTypeHint: any(named: 'mediaTypeHint'),
-          note: any(named: 'note'),
-          tag: any(named: 'tag'),
-        )).thenAnswer((_) async => createTestWishlistItem());
-
-    when(() => mockWishlistDao.findUnresolvedByText(any()))
-        .thenAnswer((_) async => null);
-
-    when(() => mockWishlistDao.updateWishlistItem(
-          any(),
-          text: any(named: 'text'),
-          mediaTypeHint: any(named: 'mediaTypeHint'),
-          clearMediaTypeHint: any(named: 'clearMediaTypeHint'),
-          note: any(named: 'note'),
-          clearNote: any(named: 'clearNote'),
-          tag: any(named: 'tag'),
-          clearTag: any(named: 'clearTag'),
-        )).thenAnswer((_) async {});
-
-    when(() => mockDb.updateItemStatus(
-          any(),
-          any(),
-          mediaType: any(named: 'mediaType'),
-        )).thenAnswer((_) async {});
-  }
-
-  group('SteamImportService', () {
-    group('importLibrary', () {
-      test('should import found games into collection', () async {
-        setupStandardMocks();
-
-        final SteamImportResult result = await sut.importLibrary(
-          apiKey: 'key',
-          steamId: '123',
-          collectionId: 1,
-          onProgress: (_) {},
-        );
-
-        expect(result.imported, 2);
-        expect(result.wishlisted, 0);
-        expect(result.updated, 0);
-        expect(result.total, 2);
-        expect(result.collectionId, 1);
-      });
-
-      test('should set inProgress for played games', () async {
-        setupStandardMocks(
-          library: <SteamOwnedGame>[
-            createTestSteamOwnedGame(
-              name: 'Played Game',
-              playtimeMinutes: 120,
-            ),
-          ],
-        );
-
-        await sut.importLibrary(
-          apiKey: 'key',
-          steamId: '123',
-          collectionId: 1,
-          onProgress: (_) {},
-        );
-
-        verify(() => mockDb.addItemToCollection(
-              collectionId: 1,
-              mediaType: MediaType.game,
-              externalId: any(named: 'externalId'),
-              platformId: 6,
-              status: ItemStatus.inProgress,
-            )).called(1);
-      });
-
-      test('should set notStarted for unplayed games', () async {
-        setupStandardMocks(
-          library: <SteamOwnedGame>[
-            createTestSteamOwnedGame(
-              name: 'Unplayed Game',
-              playtimeMinutes: 0,
-            ),
-          ],
-        );
-
-        await sut.importLibrary(
-          apiKey: 'key',
-          steamId: '123',
-          collectionId: 1,
-          onProgress: (_) {},
-        );
-
-        verify(() => mockDb.addItemToCollection(
-              collectionId: 1,
-              mediaType: MediaType.game,
-              externalId: any(named: 'externalId'),
-              platformId: 6,
-              status: ItemStatus.notStarted,
-            )).called(1);
-      });
-
-      test('should save playtime into time spent, not user comment',
-          () async {
-        setupStandardMocks(
-          library: <SteamOwnedGame>[
-            createTestSteamOwnedGame(
-              name: 'Game',
-              playtimeMinutes: 125,
-            ),
-          ],
-        );
-
-        await sut.importLibrary(
-          apiKey: 'key',
-          steamId: '123',
-          collectionId: 1,
-          onProgress: (_) {},
-        );
-
-        verify(() => mockDb.updateItemTimeSpent(100, 125)).called(1);
-        verifyNever(() => mockDb.updateItemUserComment(any(), any()));
-      });
-
-      test('should save lastPlayed as lastActivityAt', () async {
-        final DateTime lastPlayed = DateTime(2024, 1, 28);
-        setupStandardMocks(
-          library: <SteamOwnedGame>[
-            createTestSteamOwnedGame(
-              name: 'Game',
-              playtimeMinutes: 60,
-              lastPlayed: lastPlayed,
-            ),
-          ],
-        );
-
-        await sut.importLibrary(
-          apiKey: 'key',
-          steamId: '123',
-          collectionId: 1,
-          onProgress: (_) {},
-        );
-
-        verify(() => mockDb.updateItemActivityDates(
-              100,
-              lastActivityAt: lastPlayed,
-            )).called(1);
-      });
-
-      test('should not update time spent for zero playtime', () async {
-        setupStandardMocks(
-          library: <SteamOwnedGame>[
-            createTestSteamOwnedGame(
-              name: 'Game',
-              playtimeMinutes: 0,
-            ),
-          ],
-        );
-
-        await sut.importLibrary(
-          apiKey: 'key',
-          steamId: '123',
-          collectionId: 1,
-          onProgress: (_) {},
-        );
-
-        verifyNever(() => mockDb.updateItemTimeSpent(any(), any()));
-      });
-
-      test('should filter DLC and soundtracks', () async {
-        setupStandardMocks(
-          library: <SteamOwnedGame>[
-            createTestSteamOwnedGame(name: 'Real Game'),
-            createTestSteamOwnedGame(name: 'Game Soundtrack'),
-            createTestSteamOwnedGame(name: 'Game Demo'),
-          ],
-        );
-
-        final SteamImportResult result = await sut.importLibrary(
-          apiKey: 'key',
-          steamId: '123',
-          collectionId: 1,
-          onProgress: (_) {},
-        );
-
-        expect(result.total, 1);
-        expect(result.imported, 1);
-      });
-
-      test('should update duplicates instead of skipping', () async {
-        setupStandardMocks(
-          library: <SteamOwnedGame>[
-            createTestSteamOwnedGame(
-              name: 'Already Added',
-              playtimeMinutes: 300,
-              lastPlayed: DateTime(2024, 6, 15),
-            ),
-          ],
-        );
-
-        final CollectionItem collectionItem = createTestCollectionItem(
-          status: ItemStatus.notStarted,
-        );
-        when(() => mockDb.findCollectionItem(
-              collectionId: any(named: 'collectionId'),
-              mediaType: any(named: 'mediaType'),
-              externalId: any(named: 'externalId'),
-            )).thenAnswer((_) async => collectionItem);
-
-        final SteamImportResult result = await sut.importLibrary(
-          apiKey: 'key',
-          steamId: '123',
-          collectionId: 1,
-          onProgress: (_) {},
-        );
-
-        expect(result.updated, 1);
-        expect(result.imported, 0);
-        verify(() => mockDb.updateItemStatus(
-              collectionItem.id,
-              ItemStatus.inProgress,
-              mediaType: MediaType.game,
-            )).called(1);
-        verify(() => mockDb.updateItemTimeSpent(collectionItem.id, 300))
-            .called(1);
-        verifyNever(() => mockDb.updateItemUserComment(any(), any()));
-        verify(() => mockDb.updateItemActivityDates(
-              collectionItem.id,
-              lastActivityAt: DateTime(2024, 6, 15),
-            )).called(1);
-        verifyNever(() => mockDb.addItemToCollection(
-              collectionId: any(named: 'collectionId'),
-              mediaType: any(named: 'mediaType'),
-              externalId: any(named: 'externalId'),
-              platformId: any(named: 'platformId'),
-              status: any(named: 'status'),
-            ));
-      });
-
-      test('should skip time spent write when value is unchanged', () async {
-        setupStandardMocks(
-          library: <SteamOwnedGame>[
-            createTestSteamOwnedGame(
-              name: 'Already Added',
-              playtimeMinutes: 300,
-            ),
-          ],
-        );
-
-        when(() => mockDb.findCollectionItem(
-              collectionId: any(named: 'collectionId'),
-              mediaType: any(named: 'mediaType'),
-              externalId: any(named: 'externalId'),
-            )).thenAnswer((_) async => createTestCollectionItem(
-              status: ItemStatus.inProgress,
-              timeSpentMinutes: 300,
-            ));
-
-        await sut.importLibrary(
-          apiKey: 'key',
-          steamId: '123',
-          collectionId: 1,
-          onProgress: (_) {},
-        );
-
-        verifyNever(() => mockDb.updateItemTimeSpent(any(), any()));
-      });
-
-      test('should not downgrade status from completed', () async {
-        setupStandardMocks(
-          library: <SteamOwnedGame>[
-            createTestSteamOwnedGame(
-              name: 'Completed Game',
-              playtimeMinutes: 100,
-            ),
-          ],
-        );
-
-        when(() => mockDb.findCollectionItem(
-              collectionId: any(named: 'collectionId'),
-              mediaType: any(named: 'mediaType'),
-              externalId: any(named: 'externalId'),
-            )).thenAnswer((_) async => createTestCollectionItem(
-              status: ItemStatus.completed,
-            ));
-
-        await sut.importLibrary(
-          apiKey: 'key',
-          steamId: '123',
-          collectionId: 1,
-          onProgress: (_) {},
-        );
-
-        verifyNever(() => mockDb.updateItemStatus(
-              any(),
-              any(),
-              mediaType: any(named: 'mediaType'),
-            ));
-      });
-
-      test('should not downgrade status from inProgress', () async {
-        setupStandardMocks(
-          library: <SteamOwnedGame>[
-            createTestSteamOwnedGame(
-              name: 'In Progress Game',
-              playtimeMinutes: 100,
-            ),
-          ],
-        );
-
-        when(() => mockDb.findCollectionItem(
-              collectionId: any(named: 'collectionId'),
-              mediaType: any(named: 'mediaType'),
-              externalId: any(named: 'externalId'),
-            )).thenAnswer((_) async => createTestCollectionItem(
-              status: ItemStatus.inProgress,
-            ));
-
-        await sut.importLibrary(
-          apiKey: 'key',
-          steamId: '123',
-          collectionId: 1,
-          onProgress: (_) {},
-        );
-
-        verifyNever(() => mockDb.updateItemStatus(
-              any(),
-              any(),
-              mediaType: any(named: 'mediaType'),
-            ));
-      });
-
-      test('should add to wishlist when IGDB returns no match', () async {
-        setupStandardMocks(
-          library: <SteamOwnedGame>[
-            createTestSteamOwnedGame(name: 'Unknown Game'),
-          ],
-        );
-
-        when(() => mockIgdbApi.lookupSteamGames(any()))
-            .thenAnswer((_) async => <String, Game>{});
-
-        final SteamImportResult result = await sut.importLibrary(
-          apiKey: 'key',
-          steamId: '123',
-          collectionId: 1,
-          onProgress: (_) {},
-        );
-
-        expect(result.wishlisted, 1);
-        expect(result.imported, 0);
-        verify(() => mockWishlistDao.addWishlistItem(
-              text: 'Unknown Game',
-              mediaTypeHint: MediaType.game,
-              note: any(named: 'note'),
-              tag: any(named: 'tag'),
-            )).called(1);
-      });
-
-      test('should throw on empty library', () async {
-        when(() => mockSteamApi.getOwnedGames(
-              apiKey: any(named: 'apiKey'),
-              steamId: any(named: 'steamId'),
-            )).thenAnswer((_) async => <SteamOwnedGame>[]);
-
-        expect(
-          () => sut.importLibrary(
-            apiKey: 'key',
-            steamId: '123',
-            collectionId: 1,
-            onProgress: (_) {},
-          ),
-          throwsA(isA<SteamApiException>()),
-        );
-      });
-
-      test('should call onProgress callback', () async {
-        setupStandardMocks(
-          library: <SteamOwnedGame>[
-            createTestSteamOwnedGame(name: 'Game 1'),
-          ],
-        );
-
-        final List<SteamImportProgress> progressUpdates =
-            <SteamImportProgress>[];
-
-        await sut.importLibrary(
-          apiKey: 'key',
-          steamId: '123',
-          collectionId: 1,
-          onProgress: progressUpdates.add,
-        );
-
-        expect(progressUpdates.length, greaterThanOrEqualTo(3));
-        expect(
-          progressUpdates.first.stage,
-          SteamImportStage.fetchingLibrary,
-        );
-        expect(
-          progressUpdates[1].stage,
-          SteamImportStage.matchingGames,
-        );
-        expect(
-          progressUpdates.last.stage,
-          SteamImportStage.completed,
-        );
-      });
-
-      test('should match by Steam appId via lookupSteamGames', () async {
-        setupStandardMocks(
-          library: <SteamOwnedGame>[
-            createTestSteamOwnedGame(appId: 400, name: 'Portal'),
-          ],
-        );
-
-        when(() => mockIgdbApi.lookupSteamGames(any()))
-            .thenAnswer((_) async => <String, Game>{
-                  '400': const Game(id: 2, name: 'Portal'),
-                });
-
-        await sut.importLibrary(
-          apiKey: 'key',
-          steamId: '123',
-          collectionId: 1,
-          onProgress: (_) {},
-        );
-
-        verify(() => mockGameDao.upsertGame(
-              any(that: predicate<Game>((Game g) => g.id == 2)),
-            )).called(1);
-      });
-    });
-
-    group('wishlist deduplication', () {
-      test('should skip adding when unresolved wishlist item exists', () async {
-        setupStandardMocks(
-          library: <SteamOwnedGame>[
-            createTestSteamOwnedGame(name: 'Unknown Game'),
-          ],
-        );
-
-        when(() => mockIgdbApi.lookupSteamGames(any()))
-            .thenAnswer((_) async => <String, Game>{});
-
-        when(() => mockWishlistDao.findUnresolvedByText('Unknown Game'))
-            .thenAnswer((_) async => createTestWishlistItem(
-                  text: 'Unknown Game',
-                ));
-
-        await sut.importLibrary(
-          apiKey: 'key',
-          steamId: '123',
-          collectionId: 1,
-          onProgress: (_) {},
-        );
-
-        verifyNever(() => mockWishlistDao.addWishlistItem(
-              text: any(named: 'text'),
-              mediaTypeHint: any(named: 'mediaTypeHint'),
-              note: any(named: 'note'),
-              tag: any(named: 'tag'),
-            ));
-      });
-
-      test('should not touch the note of an existing wishlist row', () async {
-        setupStandardMocks(
-          library: <SteamOwnedGame>[
-            createTestSteamOwnedGame(
-              name: 'Unknown Game',
-              playtimeMinutes: 90,
-            ),
-          ],
-        );
-
-        when(() => mockIgdbApi.lookupSteamGames(any()))
-            .thenAnswer((_) async => <String, Game>{});
-
-        final WishlistItem existingItem = createTestWishlistItem(
-          id: 42,
-          text: 'Unknown Game',
-          note: 'my own note',
-        );
-        when(() => mockWishlistDao.findUnresolvedByText('Unknown Game'))
-            .thenAnswer((_) async => existingItem);
-
-        await sut.importLibrary(
-          apiKey: 'key',
-          steamId: '123',
-          collectionId: 1,
-          onProgress: (_) {},
-        );
-
-        // Only the import tag is stamped on the untagged row.
-        verify(() => mockWishlistDao.updateWishlistItem(
-              42,
-              tag: any(named: 'tag'),
-            )).called(1);
-        verifyNever(() => mockWishlistDao.addWishlistItem(
-              text: any(named: 'text'),
-              mediaTypeHint: any(named: 'mediaTypeHint'),
-              note: any(named: 'note'),
-              tag: any(named: 'tag'),
-            ));
-      });
-
-      test('should not update an existing wishlist row that already has a tag',
-          () async {
-        setupStandardMocks(
-          library: <SteamOwnedGame>[
-            createTestSteamOwnedGame(name: 'Unknown Game'),
-          ],
-        );
-
-        when(() => mockIgdbApi.lookupSteamGames(any()))
-            .thenAnswer((_) async => <String, Game>{});
-
-        when(() => mockWishlistDao.findUnresolvedByText('Unknown Game'))
-            .thenAnswer((_) async => createTestWishlistItem(
-                  id: 42,
-                  text: 'Unknown Game',
-                  tag: 'Steam 2024-01-01',
-                ));
-
-        await sut.importLibrary(
-          apiKey: 'key',
-          steamId: '123',
-          collectionId: 1,
-          onProgress: (_) {},
-        );
-
-        verifyNever(() => mockWishlistDao.updateWishlistItem(
-              any(),
-              text: any(named: 'text'),
-              mediaTypeHint: any(named: 'mediaTypeHint'),
-              clearMediaTypeHint: any(named: 'clearMediaTypeHint'),
-              note: any(named: 'note'),
-              clearNote: any(named: 'clearNote'),
-              tag: any(named: 'tag'),
-              clearTag: any(named: 'clearTag'),
-            ));
-      });
-
-      test('should create wishlist item when no duplicate exists', () async {
-        setupStandardMocks(
-          library: <SteamOwnedGame>[
-            createTestSteamOwnedGame(name: 'New Game'),
-          ],
-        );
-
-        when(() => mockIgdbApi.lookupSteamGames(any()))
-            .thenAnswer((_) async => <String, Game>{});
-
-        when(() => mockWishlistDao.findUnresolvedByText('New Game'))
-            .thenAnswer((_) async => null);
-
-        await sut.importLibrary(
-          apiKey: 'key',
-          steamId: '123',
-          collectionId: 1,
-          onProgress: (_) {},
-        );
-
-        verify(() => mockWishlistDao.addWishlistItem(
-              text: 'New Game',
-              mediaTypeHint: MediaType.game,
-              note: any(named: 'note'),
-              tag: any(named: 'tag'),
-            )).called(1);
-      });
-    });
-  });
-
-  group('SteamImportProgress', () {
-    test('should return 0 when total is 0', () {
-      const SteamImportProgress progress = SteamImportProgress(
-        stage: SteamImportStage.fetchingLibrary,
-        current: 0,
-        total: 0,
-      );
-
-      expect(progress.progress, 0.0);
-    });
-
-    test('should return fraction when total > 0', () {
-      const SteamImportProgress progress = SteamImportProgress(
-        stage: SteamImportStage.matchingGames,
-        current: 50,
-        total: 200,
-      );
-
-      expect(progress.progress, 0.25);
-    });
-  });
-
-  group('SteamImportResult', () {
-    test('should store all fields', () {
-      const SteamImportResult result = SteamImportResult(
-        imported: 10,
-        wishlisted: 5,
-        updated: 2,
-        total: 17,
-        collectionId: 42,
-      );
-
-      expect(result.imported, 10);
-      expect(result.wishlisted, 5);
-      expect(result.updated, 2);
-      expect(result.total, 17);
-      expect(result.collectionId, 42);
-    });
-  });
-}
